refactor: construct highlighted ranges already in `use-ast`
