+++ src/App.tsx
@@ -1,18 +1,18 @@
 import "./App.css";
 import { Navbar } from "./components/navbar";
 import { useExplorer } from "./hooks/use-explorer";
-import { useHighlightedRanges } from "./hooks/use-highlighted-ranges";
 import { tools } from "./lib/tools";
 import { Editor } from "./components/editor";
 import { EsquerySelectorInput } from "./components/esquery-selector-input";
 import { ToolSelector } from "./components/tool-selector";
 import { ThemeProvider } from "./components/theme-provider";
 import { Panel, PanelGroup, PanelResizeHandle } from "react-resizable-panels";
+import { useAST } from "@/hooks/use-ast";
 
 function App() {
 	const { language, tool, code, setCode, jsOptions } = useExplorer();
 
-	const highlightedRanges = useHighlightedRanges();
+	const result = useAST();
 
 	const activeTool = tools.find(({ value }) => value === tool) ?? tools[0];
 	return (
@@ -32,7 +32,11 @@ function App() {
 									)}
 									<Editor
 										value={code[language]}
-										highlightedRanges={highlightedRanges}
+										highlightedRanges={
+											!result.ok
+												? undefined
+												: result.highlightedRanges
+										}
 										onChange={value => {
 											setCode({
 												...code,
